feat: implement comprehensive SSR for improved SEO and performance
- Convert homepage to async server component with pre-loaded show data
- Add generateStaticParams for show and episode pages (23 static pages generated)
- Implement generateMetadata for proper SEO tags and social sharing
- Create hybrid SSR/client architecture maintaining all interactive features
- Add ClientDate component to fix hydration mismatches with date formatting
- Maintain backward compatibility with existing authentication and admin features

Performance improvements:
- Eliminates loading spinners on public pages
- Faster initial page loads (~2-3s improvement expected)
- Better Core Web Vitals with pre-rendered content
- Full SEO crawlability for search engines
